// src/models/enums/MovementPlane.ts
export enum MovementPlane {
  ShoulderAbduction = 'Shoulder abduction',
  ShoulderAdduction = 'Shoulder adduction',
  ShoulderFlexion = 'Shoulder flexion',
  ShoulderExtension = 'Shoulder extension',
  ShoulderRotationInternal = 'Internal shoulder rotation',
  ShoulderRotationExternal = 'External shoulder rotation',
  ShoulderRetraction = 'Shoulder retraction',
  ShoulderProtraction = 'Shoulder protraction',
  ShoulderDepression = 'Shoulder depression',
  ShoulderElevation = 'Shoulder elevation',
}